workaround for bug/behaviour in otlp-metrics-exporter (#26)
A bug in the OpenTelemetry JavaScript library related to passing an options object with undefined headers to certain exporters can lead to an application crash because of an attempt to access an undefined property. 

The reported fix should involve adding proper validation for headers in the affected exporters to prevent passing undefined values. This would prevent the crash from occurring in end-user applications.
---------

Co-authored-by: m.witti <moritz.witti@liwest.at>